Create a new fitted models from

Species are defined by an IPM which is a transition matrix from size between t and t+1, recruitment and harvest functions. Each species has these items defined for a given climate. An additionnal vector of harvest parameers is required with minimal size to harvest (dth), size above wich harvest is constant (dha).

Usage

fit_sgr(
  sv_params,
  sv_family,
  gr_params,
  gr_sigma,
  rec_params,
  species,
  max_dbh,
  delay
)

Arguments

sv_params

Named vector of survival parameters fitted for this species and climatic condition. Minimal parameters are intercept and size.

sv_family

Family list contain the details of the models used when fitting by functions such as glm. Only the inverse of the link function (linkinv) is used in the model for now. list.

gr_params

Named vector of growth parameters fitted for this species and climatic condition. Minimal parameters are intercept and size.

gr_sigma

Standard deviation of the residuals for the growth fitted model.

rec_params

Named vector of growth parameters fitted for this species and climatic condition. Minimal parameters are intercept, BATOTSP and BATOTNonSP.

species

Name of the species to run simulation on. Single char.

max_dbh

Maximum diameter of the fitted dataset. Single dbl.

delay

Delay for the species. Single dbl.