Constructor for species class

Only used in the matreex package

Usage

new_species(
  IPM,
  init_pop,
  harvest_fun,
  disturb_fun,
  harv_lim = c(dth = 175, dha = 575, hmax = 1),
  rdi_coef = NULL,
  disturb_coef = NULL
)

Arguments

IPM

ipm class object from the matreex package.

init_pop

Function to initiate the population at simulation start. Arguments must be mesh and SurfEch. Using the arguments is not mandatory, it's most useful when creating random population.

harvest_fun

Function to impact the population with harvest rule. Argument must be x, species, ...,

disturb_fun

Function to impact the population with possibles disturbances. Extra care is needed to give this function all needed parameters. Default is def_disturb.

harv_lim

Limits of harvest for a population size distribution.

dth

minimum diameter at which we cut the size distribution

dha

harvest diameter

hmax

maximum harvest rate for a size class

rdi_coef

Coefficient for RDI curve used in even harvest. The model require the intercept and slope.

disturb_coef

Species coefficient for disturbance reaction. These values and names are highly dependent on the disturbance function. BAtarget, ct and t. Should return a population state as it's take it in input, with less population than before. Unless you want zombie trees. It represent the distribution of the population to harvest