Create a new species for simulation

Species are defined by an IPM which is a transition matrix from size between t and t+1, recruitment and harvest functions (see Details). Each species has these items defined for a given climate. An additionnal vector of harvest parameters is required with minimal size to harvest (dth), size above wich harvest is constant (dha).

Usage

species(
  IPM,
  init_pop = def_init,
  harvest_fun = def_harv,
  disturb_fun = def_disturb,
  harv_lim = c(dth = 175, dha = 575, hmax = 1),
  rdi_coef = NULL,
  disturb_coef = NULL
)

Arguments

IPM

ipm class object from the matreex package.

init_pop

Function to initiate the population at simulation start. Arguments must be mesh and SurfEch. Using the arguments is not mandatory, it's most useful when creating random population.

harvest_fun

Function to impact the population with harvest rule. Argument must be x, species, ...,

disturb_fun

Function to impact the population with possibles disturbances. Extra care is needed to give this function all needed parameters. Default is def_disturb.

harv_lim

Limits of harvest for a population size distribution.

dth

minimum diameter at which we cut the size distribution

dha

harvest diameter

hmax

maximum harvest rate for a size class

rdi_coef

Coefficient for RDI curve used in even harvest. The model require the intercept and slope.

disturb_coef

Species coefficient for disturbance reaction. These values and names are highly dependent on the disturbance function. BAtarget, ct and t. Should return a population state as it's take it in input, with less population than before. Unless you want zombie trees. It represent the distribution of the population to harvest

Details

A species is defined by an IPM that is an integrated prediction matrix for growth and survival functions of the species. Since the species has other functions defined, they are accessible and editable.

init_pop

Function to initiate a new population. Default is def_init.

recruit_fun

Function that give a distribution for recruits. The default is defined from models associated with the IPM (x$IPM$rec) but it's possible to replace it. For example you can nullify the recruitment to simulate extinction.

harvest_fun

Function that give harvest density distribution when an harvest event occurs (this frequence is set at the forest scale.). The default function is def_harv with a constant harvest rate of 0.6 percent. Other functions are Uneven_harv and Even_harv.

See also